Create a datasource object. A data source provides a list of files that together comprise a dataset.

Generates an object whose @id property may be set, which then communicates with a dataset of files/objects that sooty knows about.

datasource(id = NA_character_)

dataset(...)

Arguments

id

a dataset label, see available_datasets()

...

only used by deprecated function, will become defunct

Details

The following properties are available via the @ slot:

id a dataset label, see available_datasets() (gettable and settable)

n the number of files (objects) comprising the dataset (get only)

mindate the minimum available date for the files (get only)

maxdate the maximum available date for the files (get only)

source the set of files (objects) belonging to this dataset (get only)

By default sooty maintains a local cache of the catalogue used to populate these properties. Set options("sooty.allow.cache" = FALSE) to use only the bundled sysdata, or options("sooty.cache.path" = tempdir()) to redirect the cache directory. See sooty_cache_info() for details.

Note

This was originally called dataset() which usage has now been deprecated.

Examples

op <- options("sooty.allow.cache" = FALSE)
## available dataset names
available_datasets()
#>  [1] "BREMEN-SEAICE-SMOS-north"            
#>  [2] "BREMEN-SEAICE-SMOS-south"            
#>  [3] "NSIDC_SEAICE_PS_N25km"               
#>  [4] "NSIDC_SEAICE_PS_S25km"               
#>  [5] "SEALEVEL_GLO_PHY_L4"                 
#>  [6] "antarctica-amsr2-asi-s3125-tif"      
#>  [7] "ccmp-wind-product-v2"                
#>  [8] "esacci-oc-l3s-chlor-a-merged-5day"   
#>  [9] "esacci-oc-l3s-chlor-a-merged-annual" 
#> [10] "esacci-oc-l3s-chlor-a-merged-daily"  
#> [11] "esacci-oc-l3s-chlor-a-merged-monthly"
#> [12] "ghrsst-tif"                          
#> [13] "oisst-avhrr-v02r01"                  
#> [14] "oisst-tif"                           
#> [15] "sealevel-adt-tif"                    
#> [16] "sealevel-err_sla-tif"                
#> [17] "sealevel-err_ugosa-tif"              
#> [18] "sealevel-err_vgosa-tif"              
#> [19] "sealevel-flag_ice-tif"               
#> [20] "sealevel-sla-tif"                    
#> [21] "sealevel-ugos-tif"                   
#> [22] "sealevel-ugosa-tif"                  
#> [23] "sealevel-vgos-tif"                   
#> [24] "sealevel-vgosa-tif"                  
## set to one of those
ds <- datasource("ghrsst-tif")
options(op)