Create a datasource object. A data source provides a list of files that together comprise a dataset.

Generates an object whose "@id" property may be set, which then communicates with a dataset of files/objects that sooty knows about.

datasource(id = NA_character_)

dataset(...)

Arguments

id

a dataset label, see 'datasource()@available_datasources' (get, and settable)

...

only used by deprecated function, will become defunct

Details

Compare 'curated' to 'sooty_files(curated = FALSE)', if it is curated sooty knows what dataset it belongs to, and otherwise it's just the huge list of files we're interested in for our work. All of the curation is done outside of sooty.

The following properties are available via the @ slot:

n the number of files (objects) comprising the dataset (get, not settable)

mindate the minimum available date for the files

`maxdate1 the maximum available date for the files

`source1 the set of files (objects) belonging to this dataset (get, not settable)

Note

This was originally called dataset() which usage has now been deprecated.

Examples

## available dataset names
if (interactive()) {
 available_datasets()
}
## set to one of those
ds  <- datasource("ghrsst-tif")
## access the 'ds@source' slot, files with 'date','source' (GDAL-readable)