Create a datasource object. A data source can access a list of files that together comprise a dataset.

Generates an object whose "@id" property may be set, which then communicates with a dataset of files/objects that sooty knows about.

Arguments

id

a dataset label, see 'datasorce()@available_datasources' (get, and settable)

n

the number of files (objects) comprising the dataset (get, not settable)

mindate

the minimum available date for the files

maxdate

the maximum available date for the files

source

the set of files (objects) belonging to this dataset (get, not settable)

Details

Compare 'curated' to 'sooty_files(curated = FALSE)', if its curated sooty knows what dataset it belongs to

Examples

ds <- datasource()  ## FIXME this should validate empty input
## available dataset names
available_datasets()
#>  [1] "BREMEN-SEAICE-SMOS-north"            
#>  [2] "BREMEN-SEAICE-SMOS-south"            
#>  [3] "NSIDC_SEAICE_PS_N25km"               
#>  [4] "NSIDC_SEAICE_PS_S25km"               
#>  [5] "SEALEVEL_GLO_PHY_L4"                 
#>  [6] "antarctica-amsr2-asi-s3125-tif"      
#>  [7] "ccmp-wind-product-v2"                
#>  [8] "esacci-oc-l3s-chlor-a-merged-5day"   
#>  [9] "esacci-oc-l3s-chlor-a-merged-annual" 
#> [10] "esacci-oc-l3s-chlor-a-merged-daily"  
#> [11] "esacci-oc-l3s-chlor-a-merged-monthly"
#> [12] "ghrsst-tif"                          
#> [13] "oisst-avhrr-v02r01"                  
#> [14] "oisst-tif"                           
## set to one of those
ds@id <- "oisst-tif"
ds
#> <sooty::dataset>
#>  @ id     : chr "oisst-tif"
#>  @ n      : int 14057
#>  @ mindate: POSIXct[1:1], format: "1981-09-01"
#>  @ maxdate: POSIXct[1:1], format: "2025-05-18"
#>  @ source : tibble [14,057 × 7] (S3: tbl_df/tbl/data.frame)
#>  $ date    : POSIXct[1:14057], format: "1981-09-01" "1981-09-02" ...
#>  $ source  : chr [1:14057] "/vsicurl/https://projects.pawsey.org.au/idea-oisst/1981/09/oisst-avhrr-v02r01.19810901_sst.tif" "/vsicurl/https://projects.pawsey.org.au/idea-oisst/1981/09/oisst-avhrr-v02r01.19810902_sst.tif" "/vsicurl/https://projects.pawsey.org.au/idea-oisst/1981/09/oisst-avhrr-v02r01.19810903_sst.tif" "/vsicurl/https://projects.pawsey.org.au/idea-oisst/1981/09/oisst-avhrr-v02r01.19810904_sst.tif" ...
#>  $ Dataset : chr [1:14057] "oisst-tif" "oisst-tif" "oisst-tif" "oisst-tif" ...
#>  $ Bucket  : chr [1:14057] "idea-oisst" "idea-oisst" "idea-oisst" "idea-oisst" ...
#>  $ Key     : chr [1:14057] "1981/09/oisst-avhrr-v02r01.19810901_sst.tif" "1981/09/oisst-avhrr-v02r01.19810902_sst.tif" "1981/09/oisst-avhrr-v02r01.19810903_sst.tif" "1981/09/oisst-avhrr-v02r01.19810904_sst.tif" ...
#>  $ Protocol: chr [1:14057] "/vsicurl" "/vsicurl" "/vsicurl" "/vsicurl" ...
#>  $ Host    : chr [1:14057] "https://projects.pawsey.org.au" "https://projects.pawsey.org.au" "https://projects.pawsey.org.au" "https://projects.pawsey.org.au" ...
ds@source
#> # A tibble: 14,057 × 7
#>    date                source                Dataset Bucket Key   Protocol Host 
#>    <dttm>              <chr>                 <chr>   <chr>  <chr> <chr>    <chr>
#>  1 1981-09-01 00:00:00 /vsicurl/https://pro… oisst-… idea-… 1981… /vsicurl http…
#>  2 1981-09-02 00:00:00 /vsicurl/https://pro… oisst-… idea-… 1981… /vsicurl http…
#>  3 1981-09-03 00:00:00 /vsicurl/https://pro… oisst-… idea-… 1981… /vsicurl http…
#>  4 1981-09-04 00:00:00 /vsicurl/https://pro… oisst-… idea-… 1981… /vsicurl http…
#>  5 1981-09-05 00:00:00 /vsicurl/https://pro… oisst-… idea-… 1981… /vsicurl http…
#>  6 1981-09-06 00:00:00 /vsicurl/https://pro… oisst-… idea-… 1981… /vsicurl http…
#>  7 1981-09-07 00:00:00 /vsicurl/https://pro… oisst-… idea-… 1981… /vsicurl http…
#>  8 1981-09-08 00:00:00 /vsicurl/https://pro… oisst-… idea-… 1981… /vsicurl http…
#>  9 1981-09-09 00:00:00 /vsicurl/https://pro… oisst-… idea-… 1981… /vsicurl http…
#> 10 1981-09-10 00:00:00 /vsicurl/https://pro… oisst-… idea-… 1981… /vsicurl http…
#> # ℹ 14,047 more rows